+++ src/flock/disk.h
@@ -13,7 +13,7 @@ class Grid;
 class Disk {
 public:
     Disk();
-    Disk(double centerX, double centerY, uint32_t timestamp);
+    Disk(double centerX, double centerY, uint64_t timestamp);
 
     double centerX() const { return m_centerX; }
     double centerLongitude() const { return m_centerLongitude; }
@@ -21,10 +21,10 @@ public:
     double centerY() const { return m_centerY; }
     double centerLatitude() const { return m_centerLatitude; }
 
-    uint32_t timestamp() const { return m_timestamp; }
+    uint64_t timestamp() const { return m_timestamp; }
     void addTrajectory(const Trajectory&);
-    unsigned numberOfTrajectories() const { return m_trajectories.size(); }
-    unsigned countIntersection(Disk* other) const;
+    uint32_t numberOfTrajectories() const { return m_trajectories.size(); }
+    uint32_t countIntersection(Disk* other) const;
     const std::map<uint32_t, Trajectory>& trajectories() const { return m_trajectories; }
     void dumpTrajectories() const;
     void dump() const;
@@ -34,7 +34,7 @@ private:
     double m_centerLongitude;
     double m_centerY;
     double m_centerLatitude;
-    uint32_t m_timestamp;
+    uint64_t m_timestamp;
     std::map<uint32_t, Trajectory> m_trajectories;
 };
