+++ src/gps/gpspoint.h
@@ -9,7 +9,7 @@ class GPSPoint {
 public:
     GPSPoint();
 
-    GPSPoint(double latitude, double longitude, uint32_t timestamp, uint32_t trajectoryId);
+    GPSPoint(double latitude, double longitude, uint64_t timestamp, uint32_t trajectoryId);
 
     double latitude() const { return m_latitude; }
     double latitudeMeters() const { return m_latitudeMeters; }
@@ -17,7 +17,7 @@ public:
     double longitude() const { return m_longitude; }
     double longitudeMeters() const { return m_longitudeMeters; }
 
-    uint32_t timestamp() const { return m_timestamp; }
+    uint64_t timestamp() const { return m_timestamp; }
 
     uint32_t trajectoryId() const { return m_trajectoryId; }
 
@@ -34,7 +34,7 @@ private:
     double m_latitudeMeters;
     double m_longitude;
     double m_longitudeMeters;
-    uint32_t m_timestamp;
+    uint64_t m_timestamp;
     uint32_t m_trajectoryId;
 };
